requirements.txt
appdirs==1.4.4
autopep8==2.3.1
certifi==2024.8.30
charset-normalizer==3.4.0
idna==3.10
Mako==1.3.6
MarkupSafe==3.0.2
pip==24.2
psycopg2==2.9.10
pycodestyle==2.12.1
pygubu==0.35.6
pygubu-designer==0.39.3
python-dotenv==1.0.1
requests==2.32.3
screeninfo==0.8.1
setuptools==75.1.0
tomli==2.0.2
tqdm==4.66.5
urllib3==2.2.3
